==> code/data_prep/p_00_patient_data_prepping.Rmd <==
---
title: "Patient data prepping"
author: "Yanelli Nunez"
date: "10/13/2021"
output:
 html_document:
    toc: TRUE
    toc_float: TRUE
    code_folding: hide
editor_options: 
  chunk_output_type: console
---
Objective: 

To polish the following variables:

Residency: Reduce the number of levels of the categorical variable
Place of Birth: Reduce the number of levels of the categorical variable

Note: Inside the data folder is a dictionary for the categorical variables. The variable residency is not in this dictionary. The original data contain the equivalent of zip codes as the levels in this variable and we combined those into categories using this site: 

```{r include=FALSE}

rm(list=ls())

# 1a Declare root directory
project.folder <- paste0(print(here::here()),'/')

# add folder locations
source(paste0(project.folder,'0_00_create_folder_structure.R'))

# add file locations
source(paste0(file.locations.folder,'file_locations.R'))

# load packages
 source(paste0(packages.folder,'packages.R'))

```

## 1. Load raw patient data from Johnni

The famSES variable refers to the lowest SES category of index-person and spouse. Note that the lowest value (1) refers to highest pay job title

```{r}

load(paste0(raw_data_files, "ALS_desc_1989_2013_1989_5_2k.RData"))

patient_data <- ALS_desc_1989_2013_1989_5_2k %>%
  dplyr::filter(diag %in% c("DG122", "34809", "DG122G")) %>% # select only codes for ALS and motor neurond disease
   mutate(diag = as.factor(diag),
          pattype = as.factor(pattype))

summary(patient_data)
```


## 2. Summarizing place of birth into 6 categories

Categories: 

1) Copenhagen
2) Other big cities
3) Other Denmark
4) Greenland
5) Foreign
6) Unknown


```{r}

patient_data <- patient_data %>%
mutate(place_birth = case_when(geokode >= 52 & geokode <= 62 ~ "foreign",
                                 geokode == 99 | geokode == 10 ~ "unknown",
                                 geokode > 1 & geokode <= 4 ~ "other_cities",
                                 geokode == 1 ~ "copenhagen",
                                 geokode >= 5 & geokode <= 9 ~ "other_dk",
                                 geokode == 51 ~ "greenland"),
         place_birth = as.factor(place_birth))

summary(patient_data)
```


## 3. Summarizing place of residency into 5 categoties. 

Categories:

1) Greater Copenhagen
2) Big cities of Denmark > 10,000
3) Greenland
4) Rest of Denmark
5) Unknown

***The comment out code chunk inaccurate codes the zipcode into the above categories. It actually places most observations into the rest of Denmark. 

```{r}

# classify the residency codes

#patient_data$residence <- ifelse(patient_data$postnr>=1000 & patient_data$postnr<2500, 1,          #                         ifelse(patient_data$postnr==8722|patient_data$postnr==4690|patient_data$po#stnr==2800|patient_data$postnr==8300|patient_data$postnr==2791|patient_data$postnr==9500|patient_da#ta$postnr==2640|patient_data$postnr==7600|patient_data$postnr==3060|patient_data$postnr==9700|patie#nt_data$postnr==4760|patient_data$postnr==3300|patient_data$postnr==6800|patient_data$postnr==7700|#patient_data$postnr==3500|patient_data$postnr==2625|patient_data$postnr==2665|patient_data$postnr==#8660|patient_data$postnr==5500|patient_data$postnr==3700|patient_data$postnr==4900|patient_data$pos#tnr==8500|patient_data$postnr==3600|patient_data$postnr==7430|patient_data$postnr==2680|patient_dat#a$postnr==4220|patient_data$postnr==5800|patient_data$postnr==4400|patient_data$postnr==6200|patien#t_data$postnr==4800|patient_data$postnr==3520|patient_data$postnr==3460|patient_data$postnr==2635|p#atient_data$postnr==3660|patient_data$postnr==4100|patient_data$postnr==4129|patient_data$postnr==2#600|patient_data$postnr==7800|patient_data$postnr==6100|patient_data$postnr==9400|patient_data$post#nr==9900|patient_data$postnr==4300|patient_data$postnr==9800|patient_data$postnr==6400|patient_data#$postnr==5700|patient_data$postnr==2610|patient_data$postnr==2620|patient_data$postnr==3400|patient#_data$postnr==2630|patient_data$postnr==4200|patient_data$postnr==7500|patient_data$postnr==4600|pa#tient_data$postnr==8800|patient_data$postnr==3000|patient_data$postnr==2970|patient_data$postnr==70#00|patient_data$postnr==7007|patient_data$postnr==7029|patient_data$postnr==2720|patient_data$postn#r==2750|patient_data$postnr==2920|patient_data$postnr==4700|patient_data$postnr==8600|patient_data$#postnr==4000|patient_data$postnr==7400|patient_data$postnr==7429|patient_data$postnr==7429|patient_#data$postnr==2670|patient_data$postnr==2650|patient_data$postnr==8700|patient_data$postnr==7100|pat#ient_data$postnr==7120|patient_data$postnr==6000|patient_data$postnr==8960|patient_data$postnr==890#0|patient_data$postnr==6715|patient_data$postnr==6700|patient_data$postnr==9220|patient_data$postnr#==9000|patient_data$postnr==5270|patient_data$postnr==5000|patient_data$postnr==8000|patient_data$p#ostnr==8210, 2,
#ifelse(patient_data$postnr>=3900 & patient_data$postnr<3993,3,
#ifelse(is.na(patient_data$postnr),5,4))) ) 
#
#patient_data$residence <- ifelse(is.na(patient_data$postnr),5,patient_data$residence)
#
## label categories 
#patient_data$residence<- factor(patient_data$residence,levels=c('1','2' , '3','4','5'), #labels=c("Greater Copenhagen", 
#         "Big cities of Denmark>10,000",
#         "Greenland", 
#         "Rest of Denmark",
#         "Unknown"))
#
## View summary of variable 
#
#table(patient_data$residence )
#
#summary(patient_data)
```

Corrected categorical variable for place of residence. 

Most populated cities of Denamrk
Copenhagen	1,153,615
Aarhus	237,551
Odense	145,931
Aalborg	122,219
Esbjerg 72,205
Horsens	58,646
Randers	55,780
Kolding
Vejle 

New Residence categories:

Capital area, which includes Copenhagen
Cities with >50K population  (the cities listed above other than Copenhagen)
Greenland
Rest of Denmark
Unknown

```{r}

patient_data_resid <- patient_data %>%
 mutate(residence = if_else(postnr == 3950 |
                            postnr == 3985 |
                            postnr == 3952 |
                            postnr == 3910 |
                            postnr == 3912 |
                            postnr == 3921 |
                            postnr == 3905 |
                            postnr == 3971 |
                            postnr == 3951 |
                            postnr == 3911 |
                            postnr == 3913 |
                            postnr == 3919 |  
                            postnr == 3984 |  
                            postnr == 3980 |  
                            postnr == 3930 |
                            postnr == 3982 |
                            postnr == 3923 |  
                            postnr == 3940 |  
                            postnr == 3964 |  
                            postnr == 3953 |  
                            postnr == 3992 |  
                            postnr == 3962, "Greenland",
                    if_else(postnr >= 1050 & postnr <= 4050, "Capital Region of Denmark",
                    if_else(postnr >= 5000 & postnr <= 5320, "Cities >50K population", # Odese
                    if_else(postnr >= 8000 & postnr <= 8381, "Cities >50K population", # Aarhus
                    if_else(postnr >= 9000 & postnr <= 9220, "Cities >50K population", # Aalborg
                    if_else(postnr >= 6700 & postnr <= 6715, "Cities >50K population", 
                    if_else(postnr >= 8900 & postnr < 8960,  "Cities >50K population",   
                    if_else(postnr == 7100 | postnr == 7120, "Cities >50K population", 
                    if_else(postnr == 8830, "Cities >50K population", 
                    if_else(postnr == 6000, "Cities >50K population", 
                    if_else(postnr == 8700, "Cities >50K population", "Rest of Denmark")))))))))))) %>% 
  mutate(residence = replace_na(residence, "Unknown")) %>%
  mutate(residence = as.factor(residence))


summary(patient_data_resid)                                
                                
                                  
```

## 3b. summaries
```{r}
# cases
cases <- patient_data_resid %>%
  dplyr::filter(ck == 1)

summary(cases)

# case number
count(cases)


# controls
controls <- patient_data_resid %>%
  dplyr::filter(ck == 0)

summary(controls)

#controls numbers
count(controls)


```



## 4. Make a table 1

```{r}

table1_prep <- patient_data_resid %>%
  dplyr::select(nr, lobnr, ck, sex, civst, place_birth, residence, famses) %>%
  unique() %>%
  dplyr::select(-nr, -lobnr) %>%
  mutate(ck = as.factor(if_else(ck == "0", "control", "case")),
        sex = as.factor(if_else(sex == "2", "Female", "Male")),
 famses = recode_factor(famses, 
         `1`='Group 1 (Highest)',
         `2`='Group 2',
         `3`='Group 3',
         `4`='Group 4',
         `5`='Group 5 (Lowest)',
         `9`='Group 9 (Unemployed)'),
  civst = recode_factor(civst, 
         `4`='Married',
         `3`='Divorced',
         `2`='Widower',
         `1`='Never married'),
  residence = recode_factor(residence, 
        `Capital Region of Denmark`='Capital Region of Denmark',
        `Cities >50K population`='Cities >50K population',
        `Rest of Denmark`='Rest of Denmark',
        `Greenland`='Greenland',
        `Unknown`='Unknown'),
  place_birth=recode_factor(place_birth, 
       `copenhagen`='Greater Copenhagen',
       `other_cities`='Other Major Cities',
       `other_dk`='Rest of Denmark',
       `greenland`='Greenland',
       `foreign`='Foreign',
       `unknown`='Unknown'))  
 
  
table_1 <- table1_prep %>%
  gtsummary::tbl_summary(#missing = "no",
              label = list(sex ~ "Sex",
                           famses ~ "Family SES",
                           place_birth ~ "Place of birth",
                           civst ~ "Civil status",
                           residence ~ "Place of residence"),
              by = ck,
              statistic = list(all_continuous() ~ "{mean} ({sd})",
                               all_categorical() ~ "{n} ({p}%)")) %>%
              add_overall() %>%
  bold_labels()

table_1

#table_1 %>%
 # gtsummary::as_flex_table() %>%
  #flextable::save_as_docx(path = paste0(tables.folder,'table_1_full_data.docx'))
```


## 5. Save data

```{r}
saveRDS(patient_data_resid, file = "//fsmsph/msph_ehs$/APALS_DanishData/fromJohnni/patient_data_uniform_complete.rds")
```
