==> code/data_prep/p_04_lagged_exposures.Rmd <==
---
title: "Distributed lag"
author: "Yanelli Nunez"
date: "2/5/2021"
output:
 html_document:
    toc: TRUE
    toc_float: TRUE
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

Objective:

To create yearly exposure average back to 10 years from date of diagnosis (individually for each case and control) 


```{r include=FALSE}

rm(list=ls())

# 1a Declare root directory

project.folder <- paste0(print(here::here()),'/')

# add folder locations
source(paste0(project.folder,'0_00_create_folder_structure.R'))

# add file locations
source(paste0(file.locations.folder,'file_locations.R'))

# load packages
 source(paste0(packages.folder,'packages.R'))

```

## 1. Load dataset 

Notes: 

Total cases = 4011
Total controls = 20055

1600 observations missing air pollution data (0.02% observations)

```{r}

full_dataset <- readRDS(paste0(yanelli_datasets, "joined_datasets.rds")) %>%
  dplyr::select(-year_gap, -one_year, -five_years, -ten_years, -CO_str, -EC_str, -NO2_str, -NOX_str, -O3_str, -PM10_str) %>% # remove variables we don't need
  mutate(lag_0 = diagdat - lubridate::years(1), # create date limits for each lag
         lag_1 = lag_0 - lubridate::years(1),
         lag_2 = lag_1 - lubridate::years(1),
         lag_3 = lag_2 - lubridate::years(1),
         lag_4 = lag_3 - lubridate::years(1),
         lag_5 = lag_4  - lubridate::years(1),
         lag_6 = lag_5 - lubridate::years(1),
         lag_7 = lag_6 - lubridate::years(1),
         lag_8 = lag_7 - lubridate::years(1),
         lag_9 = lag_8 - lubridate::years(1),
         lag_10 = lag_9 - lubridate::years(1)) %>%
  dplyr::select(nr, lobnr, ck, diagtype, diagdat, everything()) %>%
  drop_na(civst) %>%
  dplyr::filter(residence != "Greenland",
                residence != "Unknown") # remove greenland residents 

summary(full_dataset)

# number of cases and controls 
# cases 
total_cases <- full_dataset %>%
  dplyr::filter(ck == 1) %>%
 dplyr::group_by(nr) %>%
  summarise(count = n())

count(total_cases)

#controls
total_contrl <- full_dataset %>%
 dplyr::filter(ck == 0) %>%
  dplyr::group_by(nr, lobnr) %>%
  summarise(count = n()) %>%
  ungroup()

count(total_contrl)


```

## 2. Nest data

- Nest data by cases so we can easily select 10 years of exposure for each case/control
- The nested dataset is made up of sub-dataset. Each data subset corresponds to a patient and his/her controls
```{r}

# nest data
nested_data <- full_dataset %>% 
  mutate(nr = as.factor(nr),
         lobnr = as.factor(lobnr),
         ck = as.factor(ck)) %>%
  group_by(nr) %>%
  tidyr::nest()


### take a look in a single case and corresponding controls 
test <- as.data.frame(nested_data$data[[2]]) %>%
 dplyr::filter(date_exp <= diagdat & date_exp > lag_10) #test code to select only exposures up to 10 years before hospitalization 

summary(test)
  
```


## 3. Select exposures within the 10 year period

For each case/control, filter 10 years of exposure from date of hospitalization  

```{r}

filtered_data <- nested_data %>%
  mutate(ten_yrs = purrr::map(data, ~ dplyr::filter(.x, date_exp <= diagdat & date_exp > lag_10)),
         ten_yrs =) # select years up to lag_10
  

# explore a random dataset within the list of datasets to make sure it looks right
explore <- filtered_data$ten_yrs[[2]]
summary(explore)


# unnest dataset for 10 years 
ten_year <- filtered_data %>%
  dplyr::select(-data) %>%
  unnest(., cols = ten_yrs)


```

## 4. Lag periods

- I create a variable lags that labels each exposure with the lag number they belong to
- I checked how many exposure estimates each lag has
- Exposure estimates are per month so we set the requirement that each lag should at estimates for at least 9  months. After removing cases with < 9 exposure estimates, we ended up with *3825 cases* and *19121 controls* (~4% cases removed)

```{r}

lags <- ten_year %>%
  mutate(lags = if_else(date_exp < diagdat & date_exp >= lag_0, 0,
                if_else(date_exp < lag_0 & date_exp >= lag_1, 1, 
                if_else(date_exp < lag_1 & date_exp >= lag_2, 2,
                if_else(date_exp < lag_2 & date_exp >= lag_3, 3,
                if_else(date_exp < lag_3 & date_exp >= lag_4, 4,
                if_else(date_exp < lag_4 & date_exp >= lag_5, 5,
                if_else(date_exp < lag_5 & date_exp >= lag_6, 6,
                if_else(date_exp < lag_6 & date_exp >= lag_7, 7,
                if_else(date_exp < lag_7 & date_exp >= lag_8, 8,
                if_else(date_exp < lag_8 & date_exp >= lag_9, 9, 10))))))))))) %>%
  dplyr::select(-lag_0, -lag_1, -lag_2, -lag_3, -lag_4, -lag_5, -lag_6, -lag_7, -lag_8, -lag_9, -lag_10) %>%
  group_by(nr, lobnr, lags) %>%
  mutate(mean_pm2.5 = mean(PM25_str),
         months_lag = n()) %>%  # check how many observations each lag has
ungroup()

# identify ID of cases and controls that have lags with fewer than  9 exposure estimates
labels_incomp <- lags %>%
dplyr::filter(months_lag < 9) %>%
dplyr::select(nr, lobnr) %>%
 dplyr::distinct()

# remove  cases and controls with lags that have miss estimates

lags_filt <- anti_join(lags, labels_incomp)


# check how many cases and controls we end up with

# cases 
lag_cases <- lags_filt %>%
  dplyr::filter(ck == 1) %>%
 dplyr::group_by(nr) %>%
  summarise(count = n())

count(lag_cases)

#controls
lag_contrl <- lags_filt %>%
 dplyr::filter(ck == 0) %>%
  dplyr::group_by(nr, lobnr) %>%
  summarise(count = n()) %>%
  ungroup()

count(lag_contrl)

```



## 5. Clean & criteria selection

- Removed variables we don't need and keep only the lag averages (remove monthly estimates)
- Kept only the cases/controls that have exposure estimates for all 11 lags (0-10)
- Kept only cases that have at least one control and controls that have an associated case

Notes:
We have *18265* complete controls 
We have *3801* complete cases 

All case have at least 2 controls

```{r}

# keep only the average estimates and remove the monthly estimates
clean_lag <- lags_filt %>%
  ungroup() %>%
  dplyr::select(-PM25_str, -date_exp, -months_lag) %>%
  dplyr::distinct()

summary(clean_lag)

# keep only cases/controls that have all the 11 lags (0-10 lag) and cases that have at least one control
complete_lags <- clean_lag %>%
  dplyr:: group_by(nr,lobnr) %>%
 dplyr::mutate(counts = n()) %>%
  dplyr::filter(counts == 11) %>%
  dplyr::select(-counts) %>%
  ungroup() %>%
  group_by(nr) %>%
  dplyr::mutate(case_control = length(unique(ck))) %>% # ID to check that each case has at least one control and that each control has a case
  dplyr::filter(case_control == 2) %>%
  dplyr::select(-case_control)

summary(complete_lags)

# check how many cases and controls we ended up with
num_controls <- complete_lags %>%
dplyr::filter(ck == 0) %>% # selecting controls
 group_by(nr) %>% 
  count(lobnr) %>%
  ungroup()

count(num_controls)

num_cases <- complete_lags %>%
dplyr::filter(ck == 1) %>% # selecting only cases
 count(nr) %>%
  ungroup()

count(num_cases)


```

## 6. Save data 

```{r}
 write_csv(complete_lags, paste0(yanelli_datasets, "lagged_10yrs.csv"))
```
